git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@4862 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ regress/lib/libgfarm/gfarm/gfs_pio_open/file_trunc_not_writable.sh
@@ -4,6 +4,10 @@
 
 gfs_pio_test=$testbin/../gfs_pio_test/gfs_pio_test
 
+if $regress/bin/am_I_gfarmroot; then
+  exit $exit_unsupported
+fi
+
 trap 'gfrm -f $gftmp; rm -f $localtmp; exit $exit_trap' $trap_sigs
 
 if gfreg $data/1byte $gftmp && gfchmod 0444 $gftmp
